==> scripts/1_cortisol/manuscript/1_Demographics_summaries_prep_cort_data.Rmd <==
---
title: "getting data ready for longitudinal cort analysis"
author: "Michelle.VanTieghem"
date: "March 27, 2018"
output:
  html_document:
    number_sections: no
    toc: yes
    toc_float:
      collapsed: no
      smooth_scroll: yes
  pdf_document:
    toc: yes
---
# Notes
Data cleaning and demographics per wave 

```{r, warnings = F, include = F, message = F}
source("../../0_R_analysis_setup_file.R")
```


```{r, warnings = F, message = F}

load(file = "../../../data/1_cort_data/1D_long_cort_testos_cleaned.Rdata")

nrow(cort_merge8)
sum(is.na(cort_merge8$batch))
sum(is.na(cort_merge8$cort_values_fixed_no_outliers))

summary(cort_merge8$batch) # make sure this exists! added in prior step with testosterone. 
```

## include/ exclude - double check age ranges for comps
```{r, warnings = F}
# check that comp age range is 4+ and PI keeping the 3.9 
cort_merge8 %>%
  group_by(GROUP) %>%
  dplyr::summarize(min_age = min(corrected_cort_age_yrs, na.rm = T), 
            max_age = max(corrected_cort_age_yrs, na.rm = T))

```


# remove extreme value and subject with no batch for cort. 
```{r}
cort_merge8 <- cort_merge8 %>%
  filter(cort_values_fixed_no_outliers < 75 & !is.na(batch))
```

## fixing variables 
```{r, warnings = F}
# these need to be re-leveled so that waking = first 
cort_merge8$index3.f <- factor(cort_merge8$index3.f, 
                            levels = c("waking", "afternoon","evening"))
# order the levels!!! 
cort_merge8$time_of_day <- factor(cort_merge8$time_of_day,levels = c("waking", "evening"))
# code this so wave 1 = 0 and then all positive nums are change 
cort_merge8$wave_num <- as.numeric(ifelse(cort_merge8$index_wave == 1, 0, 
                                       ifelse(cort_merge8$index_wave == 2, 2, 4)))
sublist <- (unique(cort_merge8$IDENT_SUBID)) 
length(sublist)
# need to choose to one and stick to it! 
#more about waking or slope? 
#names(cort_merge8)[grepl("RCADS", names(cort_merge8))]
nrow(cort_merge8)
# fix gender variable
cort_merge8$sex <- (cort_merge8$DEM_3_GENDER_CHILD)
```

## get info on dataset of cort sampling 
```{r, warnings = F}
# table for subjects with cort 
table <-cort_merge8 %>% 
    group_by(GROUP, index_wave, index_day, index4, time_of_day) %>% 
  # sort data by group and wave 
        dplyr::summarize(N= n(), 
                   mean_PDS = mean(PDS_mean, na.rm = T),
                   mean_age = mean(corrected_cort_age_yrs, na.rm = TRUE),  
                   min_age = min(corrected_cort_age_yrs, na.rm = TRUE), 
                              max_age = max(corrected_cort_age_yrs, na.rm = TRUE),
                   prop_female = mean(sex, na.rm = TRUE))
  
table
```

## plot age across waves for each group 
```{r}
# get min age for each subject. 
min_age <- cort_merge8 %>%
  group_by(IDENT_SUBID) %>%
 dplyr:: summarize(min_age = min(corrected_cort_age_yrs))
for (i in 1:nrow(cort_merge8)){
  sub <-cort_merge8$IDENT_SUBID[i]
  cort_merge8$min_age[i] <- min_age$min_age[min_age$IDENT_SUBID== sub]
}
```

```{r}
AgebyParticipant_Cort <-ggplot(cort_merge8,
                         aes(colour=GROUP,
                             y=fct_reorder(IDENT_SUBID, min_age), x=corrected_cort_age_yrs))+
  geom_line(aes(group = IDENT_SUBID), size=0.5, alpha = 0.5) + 
  ylab("Participants") + xlab("Age") +
  geom_point(size=1,  alpha = 0.2) + 
  theme_classic() + my_colors + facet_grid (~ GROUP) +
  theme(axis.line = element_line(colour = "black"),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        legend.position="none",
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank()) + 
   scale_x_continuous(breaks = seq(4,21, by = 3), limits=c(4, 21)) 
pdf(file = "figures/Manuscript_Figure1_AgebyParticipant_cort.pdf", width = 6, height =4)
AgebyParticipant_Cort
dev.off()
#AgebyParticipant_Cort
```



## create a separate raw df for plotting that averages across DAYS.
```{r}
summary(cort_merge8$GROUP)
cort_merge8$GROUP.n <- with(cort_merge8, ifelse(GROUP == "PI", 1, 0))
summary(as.factor(cort_merge8$GROUP.n))
cort_ave_days_m <- cort_merge8 %>%
  group_by(IDENT_SUBID, index_wave, time_of_day) %>%
  # this will provide average cort by each wave, and time of day...
  #aka average across 2 days if that subject has 2 days of data!
  dplyr::summarize(n_cort_samples = n(), 
            cort_values_fixed_no_outliers = mean(cort_values_fixed_no_outliers),
            PDS_mean = mean(PDS_mean),
            corrected_cort_age_yrs = mean(corrected_cort_age_yrs), 
            GROUP.n = mean(GROUP.n), 
            sex = mean(sex))
#nrow(cort_ave_days_m)
cort_ave_days_m$GROUP <- as.factor(ifelse(cort_ave_days_m$GROUP.n == 0, "COMP", "PI"))
summary(cort_ave_days_m$GROUP)

```

# Quality Control info 

## missing samples or time-stamps across the day 

```{r}
# calculate number of cort samples per day, per wave, per subject. 
cort_count_m <- cort_merge8 %>%
  mutate(timing_exists = ifelse(!is.na(time.decimal.fixed), 1, 0)) %>%
  group_by(IDENT_SUBID, index_wave,  index_day) %>%
  dplyr::summarize(n_cort_samples = n(), 
             full_timing = sum(timing_exists), 
            corrected_cort_age_yrs = mean(corrected_cort_age_yrs), 
            GROUP = mean(GROUP.n), 
            sex = mean(sex))

# visualize, 
hist(cort_count_m$n_cort_samples)

# make a table to see how many days missing 1 + samples, by group 
cort_missing_table <- cort_count_m %>%
  # mark days of cort with full set (4) or missing sample(s)
  mutate(full_set = ifelse(n_cort_samples == 4 & 
                             full_timing == 4, 1, 0)) %>%
  group_by(full_set) %>%
  dplyr::summarize(n_days = n())
cort_missing_table
total_days <- cort_missing_table$n_days[1] + cort_missing_table$n_days[2]
prop_days_missing_samples <- cort_missing_table$n_days[cort_missing_table$full_set==0]/ total_days 
# 21 % of diurnal datasets (days) are missing at least 1 sample or time stamp!  
prop_days_missing_samples


```

## morning sample interval timing
```{r}
morning_timing <- cort_merge8 %>%
  select(IDENT_SUBID, time.decimal.fixed, 
         time_of_day, index4, index_day, index_wave, GROUP) %>%
  filter(time_of_day == "waking") 

morning_timing <- morning_timing %>% 
  spread(index4, time.decimal.fixed) %>%
  rename(morning = '1', 
         morning45 = '2') %>%
  mutate(time_diff = morning45 - morning) %>%
  filter(!is.na(time_diff))

#nrow(morning_timing)

#hist(morning_timing$time_diff)
  
morning_timing_table <- morning_timing %>%
  group_by(IDENT_SUBID, index_day, index_wave) %>%
  # mark if longer than 45 minutes different.
  mutate(too_long = ifelse(time_diff > 0.75, 1, 0)) %>%
  dplyr::summarize(n_too_long = sum(too_long))

nrow(morning_timing_table)
 # 24 % of samples were too long.  
sum(morning_timing_table$n_too_long)/nrow(morning_timing_table)
```

## correlations between day1-day2
```{r}

cort_df <- cort_merge8 %>%
  select(IDENT_SUBID, GROUP,time_of_day, index_day, index_wave, index4,
         cort_values_fixed_no_outliers) %>%
  # make wide version so that day 1 and day 2 are side by side
  spread(index_day, cort_values_fixed_no_outliers) %>%
  rename(day1 = '-1', day2 = '1') 

# morning samples correlated day1-day2
morning1_df <- cort_df %>%
  filter(index4 == 1)
cor.test(morning1_df$day1, morning1_df$day2)

morning2_df <- cort_df %>%
  filter(index4 == 2)
cor.test(morning2_df$day1, morning2_df$day2)


# evening samples correlated day1-day2
evening5_df <- cort_df %>%
  filter(index4 == 3)
cor.test(evening5_df$day1, evening5_df$day2)

evening8_df <- cort_df %>%
  filter(index4 == 4)
cor.test(evening8_df$day1, evening8_df$day2)


```

# CONFOUND T-TESTS 

## t-test table for covariates - group differences
```{r}

cort_by_sub <- cort_merge8 %>%
  group_by(IDENT_SUBID) %>%
  dplyr::summarize(GROUP = GROUP[1],
                   sex = mean(sex, na.rm = T), 
                   corrected_cort_age_yrs = mean(corrected_cort_age_yrs, na.rm = T), 
                   batch= mean(batch, na.rm = T), 
                   meds = mean(as.numeric(as.character(meds)), na.rm = T))
# set up x andy variables 
my_dataX <- data.frame(with(cort_by_sub, cbind( GROUP)))
my_dataY <- data.frame(with(cort_by_sub, cbind(sex, corrected_cort_age_yrs, batch, meds)))

# make an empty table.
compiled_ttests <- data.frame(x = NA, y = NA, tvalue = NA, df = NA, pvalue = NA)

# set up loop for making t-tests 
# for each column in mydata...
for (i in 1:ncol(my_dataX)){
  print (i)
  var1 <- my_dataX[i]
  var1_name <- names(var1)
     var1_test <- as.matrix(var1)
  # get every other column in the dataframe...
  for (j in 1:ncol(my_dataY)){
   # print (j)
     var2 <-my_dataY[j]
     var2_name <- names(var2)
     var2_test <- as.matrix(var2)
     # run the correlation and save the p-value...
     t_value <- t.test(var2_test ~ var1_test)[1]
     d_f <- t.test(var2_test ~ var1_test)[2]
     p_value <- t.test(var2_test ~ var1_test)[3]
     # keep the p-value 
    row <- cbind(x=var1_name, y=var2_name, tvalue = t_value, df = d_f, pvalue = p_value)
    compiled_ttests <- rbind(compiled_ttests, row)

  }
}
compiled_ttests$sig_level <- ifelse(compiled_ttests$pvalue < 0.001, '***', 
                                          ifelse(compiled_ttests$pvalue < 0.01, '**',
                                                 ifelse(compiled_ttests$pvalue < 0.05, '*', " ")))
compiled_ttests$x <- unlist(compiled_ttests$x)
compiled_ttests$y <- unlist(compiled_ttests$y)
compiled_ttests$tvalue <- unlist(compiled_ttests$tvalue)
compiled_ttests$df <- unlist(compiled_ttests$df)
compiled_ttests$pvalue <- unlist(compiled_ttests$pvalue)

compiled_ttests_cort <- compiled_ttests 
save(compiled_ttests_cort, file = "tables/confound_variable_t-tests_with_pvalues.Rdata")
```

# BASIC DEMOGRAPHICS ON THESE SUBS. 
## by subject
```{r}
library(readr)
# also get min index20 for each subject 
cort_merge8$index20.n <- parse_number(cort_merge8$index20)
# get min age for each subject. 
cort_merge8$min_index20 <- NA
min_index20_df <- cort_merge8 %>%
  group_by(IDENT_SUBID) %>%
 dplyr:: summarize(min_index20 = min(index20.n, na.rm = T))
head(min_index20_df)
for (i in 1:nrow(cort_merge8)){
  sub <- cort_merge8$IDENT_SUBID[i]
 # print(sub)
  cort_merge8$min_index20[i] <- min_index20_df$min_index20[min_index20_df$IDENT_SUBID== sub]
}
head(cort_merge8$index20.n)
head(cort_merge8$min_index20)
Cortisol_demographic_table <- cort_merge8 %>% 
  # get only 1 row per subj
  filter((corrected_cort_age_yrs == min_age) & (index20.n == min_index20)) %>%
  filter(!is.na(batch)) %>%
  # filter by group
  group_by(GROUP) %>%
  # get basic demo info 
  dplyr::summarize(n = n(), 
          mean_age_wave1 = round(mean(corrected_cort_age_yrs, na.rm = T), 2),
            min_age_wave1 = round(min(corrected_cort_age_yrs), 2),
            max_age_wave1 =round(max(corrected_cort_age_yrs), 2),
            proportion_female = round(mean(sex), 2), 
          N_female = sum(sex))
Cortisol_demographic_table
save(Cortisol_demographic_table, file = "tables/Cortisol_demographic_table.Rdata")
```

## adoption specific demo
```{r}
# merge in adoption demographics from master. 
master <- read.csv("../../../data/master/J.Data_Entry_Master_8_17_2017_fixed_fixed.csv")
adopt_var <- master[grepl("IAI_", names(master))]
adopt_df <- cbind(master$IDENT_SUBID, adopt_var)
head(adopt_df)
names(adopt_df)[1] <- "IDENT_SUBID"
summary(cort_merge8$GROUP.n)
cort_merge8 <- merge(cort_merge8, adopt_df, by = "IDENT_SUBID")
summary(cort_merge8$GROUP.n)
# PI only demo 
PI_adopt_vars <- cort_merge8 %>%
  filter(corrected_cort_age_yrs == min_age & GROUP == "PI") %>% 
  dplyr::summarize(median_age_orph = median(IAI_3_AGEORPH, na.rm = T), 
            mean_age_orph = mean(IAI_3_AGEORPH, na.rm = T), 
            min_age_orph = min(IAI_3_AGEORPH, na.rm = T),
            max_age_orph = max (IAI_3_AGEORPH, na.rm = T),
            median_age_adopt = median(IAI_7_AGEADOPT, na.rm = T),
            mean_age_adopt = mean(IAI_7_AGEADOPT, na.rm = T),
           min_age_adopt = min(IAI_7_AGEADOPT, na.rm = T),
           max_age_adopt = max(IAI_7_AGEADOPT, na.rm = T))
#PI_adopt_vars
write.csv(PI_adopt_vars, "tables/PI_adopt_vars_in_Cortisol_Sample.Rdata")
```

## by wave
```{r}
cort_ave_days_m_time <- cort_merge8 %>%
  group_by(IDENT_SUBID, index_wave) %>%
  # this will provide average cort by each wave, and time of day...
  #aka average across 2 days if that subject has 2 days of data!
  dplyr::summarize(cort_values_fixed_no_outliers = mean(cort_values_fixed_no_outliers),
            PDS_mean = mean(PDS_mean),
            corrected_cort_age_yrs = mean(corrected_cort_age_yrs), 
            GROUP.n = mean(GROUP.n), 
            sex = mean(sex), 
            batch = mean(batch, na.rm = T))
summary(as.factor(cort_merge8$GROUP))
summary(as.factor(cort_ave_days_m_time$GROUP.n))
nrow(cort_ave_days_m_time)
test <- subset(cort_ave_days_m_time, IDENT_SUBID == "SB330")
nrow(test)
cort_ave_days_m_time$GROUP <- as.factor(ifelse(cort_ave_days_m_time$GROUP.n == 0, "COMP", "PI"))
summary(cort_ave_days_m_time$GROUP)
cortisol_demo_table_by_wave <- cort_ave_days_m_time %>%
  filter(!is.na(batch)) %>%
  group_by( GROUP, index_wave) %>%
  dplyr::summarize(N = n(),
            mean_age = round(mean(corrected_cort_age_yrs, na.rm = T), 2),
            sd_age = round(sd(corrected_cort_age_yrs, na.rm = T), 2),
            min_age = round(min(corrected_cort_age_yrs), 2),
            max_age=round(max(corrected_cort_age_yrs), 2),
            proportion_female = round(mean(sex), 2), 
            N_female = sum(sex))
cortisol_demo_table_by_wave

save(cortisol_demo_table_by_wave, file = "tables/Cortisol_demographic_table_by_wave.Rdata")
cort_ave_days_m$sex <- as.factor(ifelse(cort_ave_days_m$sex == 0, "male", "female"))
```


# plot raw data
raw data plot with loess line smoothing
```{r, warnings = F}
ggplot(cort_ave_days_m, aes(x = corrected_cort_age_yrs, y = cort_values_fixed_no_outliers, 
                     color = as.factor(GROUP))) + facet_grid (~ time_of_day) + 
  geom_point (alpha = 0.5, size = 0.5) + 
  ylab ("Raw cortisol values (averaged across days)") +
 geom_line(aes(group = IDENT_SUBID), alpha = 0.5, size = 0.5) + 
  geom_smooth() + theme_classic() +   my_colors 
```

## make centered vars
```{r}
cort_merge8$sex.centered <- cort_merge8$sex- mean(cort_merge8$sex)
cort_merge8$index_wave.c <- as.numeric(with(cort_merge8, ifelse(index_wave == 1, -1, ifelse(index_wave == 2, 0, 3))))
cort_merge8$index_day.c <- as.numeric(cort_merge8$index_day)
cort_merge8$batch.c <-as.numeric(as.character(cort_merge8$batch)) -  mean(as.numeric(as.character(cort_merge8$batch)), na.rm = T)
nrow(cort_merge8)
length(unique(cort_merge8$IDENT_SUBID))
```

## save this data 
```{r}
save(cort_merge8, file = "../../../data/1_cort_data/Cortisol_Ready_to_analyze.Rdata")
```
